refactor :: 아이콘 svg 최적화
