+++ src/components/icons/IconCamera.tsx
@@ -5,8 +5,8 @@ const IconCamera = () => {
   return (
     <Svg width="17" height="16" viewBox="0 0 17 16" fill="none">
       <Path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
+        fillRule="evenodd"
+        clipRule="evenodd"
         d="M4.63337 4L5.66004 2.28667C5.76004 2.12667 5.98671 2 6.16004 2H10.84C11.02 2 11.24 2.12667 11.34 2.28667L12.3667 4H13.8334C14.5667 4 15.1667 4.59333 15.1667 5.33333V12C15.1667 12.74 14.5734 13.3333 13.8334 13.3333H3.16671C2.43337 13.3333 1.83337 12.74 1.83337 12V5.33333C1.83337 4.6 2.42671 4 3.16671 4H4.63337ZM7.8335 3.80001C7.53895 3.80001 7.30017 4.03879 7.30017 4.33334C7.30017 4.62789 7.53895 4.86667 7.8335 4.86667H9.16684C9.46139 4.86667 9.70017 4.62789 9.70017 4.33334C9.70017 4.03879 9.46139 3.80001 9.16684 3.80001H7.8335ZM6.70017 8.66667C6.70017 7.67256 7.50606 6.86667 8.50017 6.86667C9.49428 6.86667 10.3002 7.67256 10.3002 8.66667C10.3002 9.66079 9.49428 10.4667 8.50017 10.4667C7.50606 10.4667 6.70017 9.66079 6.70017 8.66667ZM8.50017 5.80001C6.91695 5.80001 5.6335 7.08346 5.6335 8.66667C5.6335 10.2499 6.91695 11.5333 8.50017 11.5333C10.0834 11.5333 11.3668 10.2499 11.3668 8.66667C11.3668 7.08346 10.0834 5.80001 8.50017 5.80001Z"
         fill="#8C8C8C"
       />
