+++ src/types/icon.type.ts
@@ -1,0 +1,6 @@
+interface IconType {
+  onPress?: () => void;
+  color?: string;
+}
+
+export default IconType;
