+++ src/types/index.ts
@@ -1,0 +1,1 @@
+export type {default as IconType} from './icon.type';
